Added Last Will and Testament to sketch and Home Assistant sample config
Last Will and Testament (LWT) notifies the MQTT broker that the device is online.  If the device goes offline, the broker will detect that the device is unavailable (after a timeout period) and show that the device is offline.

When a device in Home Assistant is configured with an availability_topic, the device's status will be shown as unavailable if the device goes offline.

I did not add the availability_topic to the 'cats last fed mqtt' topic so it's status will not be affected if the device goes offline.  The 'catfeeder remaining' sensor can be used as a trigger to get a notification that the feeder is offline.